WIP on restructuring shared imports
