+++ packages/shared/package.json
@@ -3,11 +3,12 @@
 	"version": "1.0.0",
 	"license": "AGPL-3.0-only",
 	"type": "module",
+	"main": "src/index.ts",
 	"scripts": {
 		"test": "ts-mocha -p tsconfig.json src/**/*.spec.ts",
 		"test:build": "tsc --noEmit",
 		"test:debug": "ts-mocha -p tsconfig.json src/**/*.spec.ts --inspect-brk",
-		"test:watch": "ts-mocha -p tsconfig.json --reporter min --watch --watch-files src/**/*.spec.ts src/**/*.spec.ts",
+		"test:watch": "ts-mocha -p tsconfig.json --reporter min --watch-files src/**/*.ts  --watch src/**/*.spec.ts",
 		"build": "tsc --build"
 	},
 	"dependencies": {
