+++ packages/shared/src/parse/scraper.ts
@@ -1,7 +1,7 @@
 import cheerio, { Node, CheerioAPI } from "cheerio";
 import axios from "axios";
-import { ParsedNode } from "packages/shared/src/schemas/scrape";
-import { ScrapeResult } from "./schemas/scrape";
+import { ParsedNode } from "@lawbrador/shared";
+import { ScrapeResult } from "@lawbrador/shared";
 
 interface NodeAttributes {
   id?: string;
