+++ packages/shared/src/restructure/selectRuleSet.ts
@@ -1,0 +1,20 @@
+import {DocumentRuleSet, getFirstMatching, ParsedNode, SourceSiteConfig} from "..";
+
+export function selectRuleSet(
+  root: ParsedNode,
+  config: SourceSiteConfig
+): DocumentRuleSet | null {
+  for (let i = 0; i < config.documentRuleSets.length; i++) {
+    const ruleSet = config.documentRuleSets[i];
+    debugger;
+    const matchesConditions =
+      //@ts-ignore - this somehow breaks a recursivity-check in TS
+      ruleSet.conditionRules?.every((rs) => getFirstMatching(root, rs)) ?? true;
+    const matchesBody = !!getFirstMatching(root, ruleSet.bodyRule);
+    if (matchesConditions && matchesBody) {
+      return ruleSet;
+    }
+  }
+  return null;
+}
+
